test: move some test files
