+++ src/tests/output/entry/years.test.ts
@@ -1,6 +1,6 @@
-import { newPost } from '../../post'
-import { getDate, getEntry, getUrl } from '../../output/entry'
-import { Format } from '../../output/format'
+import { newPost } from '../../../post'
+import { getDate, getEntry, getUrl } from '../../../output/entry'
+import { Format } from '../../../output/format'
 
 describe('post entry format', () => {
     const path = 'test/samples/posts/blocking-qobjects-signals.md'
